removing duplicated code -- whisper_online_vac
